Reorganized processor and service components into dedicated namespaces
