--- source/LootDumpProcessor/Process/IKeyGenerator.cs
@@ -1,6 +1,0 @@
-﻿namespace LootDumpProcessor.Process;
-
-public interface IKeyGenerator
-{
-    string Generate();
-}
