+++ source/LootDumpProcessor/Process/Processor/FileProcessor/FileProcessor.cs
@@ -1,7 +1,7 @@
 using LootDumpProcessor.Model;
 using LootDumpProcessor.Model.Processing;
-using LootDumpProcessor.Process.Processor.v2.LooseLootProcessor;
-using LootDumpProcessor.Process.Processor.v2.StaticLootProcessor;
+using LootDumpProcessor.Process.Processor.LooseLootProcessor;
+using LootDumpProcessor.Process.Processor.StaticLootProcessor;
 using LootDumpProcessor.Storage;
 using Microsoft.Extensions.Logging;
