+++ source/LootDumpProcessor/Process/Services/KeyGenerator/IKeyGenerator.cs
@@ -1,0 +1,6 @@
+﻿namespace LootDumpProcessor.Process.Services.KeyGenerator;
+
+public interface IKeyGenerator
+{
+    string Generate();
+}
