+++ source/LootDumpProcessor/ServiceCollectionExtensions.cs
@@ -1,18 +1,19 @@
-﻿using System.Collections.Frozen;
-using LootDumpProcessor.Model.Config;
-using LootDumpProcessor.Model.Output.StaticContainer;
+﻿using LootDumpProcessor.Model.Config;
 using LootDumpProcessor.Process;
 using LootDumpProcessor.Process.Collector;
+using LootDumpProcessor.Process.Processor.AmmoProcessor;
 using LootDumpProcessor.Process.Processor.DumpProcessor;
 using LootDumpProcessor.Process.Processor.FileProcessor;
-using LootDumpProcessor.Process.Processor.v2.AmmoProcessor;
-using LootDumpProcessor.Process.Processor.v2.LooseLootProcessor;
-using LootDumpProcessor.Process.Processor.v2.StaticContainersProcessor;
-using LootDumpProcessor.Process.Processor.v2.StaticLootProcessor;
+using LootDumpProcessor.Process.Processor.LooseLootProcessor;
+using LootDumpProcessor.Process.Processor.StaticContainersProcessor;
+using LootDumpProcessor.Process.Processor.StaticLootProcessor;
 using LootDumpProcessor.Process.Reader.Filters;
 using LootDumpProcessor.Process.Reader.Intake;
+using LootDumpProcessor.Process.Services.ComposedKeyGenerator;
+using LootDumpProcessor.Process.Services.ForcedItemsProvider;
+using LootDumpProcessor.Process.Services.KeyGenerator;
+using LootDumpProcessor.Process.Services.TarkovItemsProvider;
 using LootDumpProcessor.Process.Writer;
-using LootDumpProcessor.Serializers.Yaml;
 using LootDumpProcessor.Storage;
 using LootDumpProcessor.Storage.Implementations.File;
 using LootDumpProcessor.Storage.Implementations.Memory;
